Fix: Repair delete record from user methods.
- Specify a new delete function in the repository instead of using crudRepository delete.
- Modify entities to not allow user_id to be null for renovationRecords and add orphanRemoval to the list of records in the user.
- Return different responses based on the current user and record id given to delete.

+++ src/main/java/nz/ac/canterbury/seng302/homehelper/controller/RenovationController.java
@@ -130,11 +130,19 @@ public class RenovationController {
     /**
      * Deletes renovation record by its id, redirects back to my records page
      * @param id of the record to be deleted
-     * @return redirect to my records page
+     * @return response based on whether the record id exists, if the user doesn't have permission to delete the record, or
+     * if the deletion was successful
      */
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<Void> deleteRecord(@PathVariable Long id) {
         logger.info("DELETE /renovations/");
+        RenovationRecord record = renovationRecordService.getRecordById(id);
+        if (record == null) {
+            return ResponseEntity.notFound().build();
+        }
+        if (!loginService.getUserByEmail().equals(record.getUser())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         renovationRecordService.removeRenovationRecord(id);
         return ResponseEntity.noContent().build();
     }
